requirements.txt
# 泰摸鱼吧 - 核心依赖包 (简化版)

# 核心框架
Flask==3.0.3
Jinja2==3.1.4
Werkzeug==3.0.3

# 数据库
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30

# 数据库驱动
PyMySQL==1.1.1
pymssql==2.2.11
pyodbc==5.1.0
# cx-Oracle==8.3.0  # 需要Oracle客户端，暂时注释

# 缓存
redis==5.0.7
Flask-Caching==2.1.0

# 异步任务
Celery==5.3.6
kombu==5.3.6

# 认证和安全
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
# Flask-Security==5.6.2  # 暂时注释，避免依赖冲突

# 表单处理
Flask-WTF==1.2.1
WTForms==3.1.2

# 权限管理
Flask-Principal==0.4.0

# 管理界面
Flask-Admin==1.6.1

# API相关
Flask-RESTful==0.3.10
Flask-CORS==4.0.0

# 配置管理
python-dotenv==1.0.1

# 测试
pytest==8.2.2
pytest-flask==1.3.0

# 开发工具
gunicorn==22.0.0

# 其他工具
requests==2.32.3
python-dateutil==2.9.0